Enum wasi_test_generator::wasi_version::WasiVersion

pub enum WasiVersion {
    Unstable,
    Snapshot1,
    Nightly_2022_10_18,
}

Variants

Unstable

A.K.A. Snapshot0

Snapshot1

Nightly_2022_10_18

This is for making tests pass on Apple M1 while still keeping the old test for compatibility reasons

Implementations

impl WasiVersion

pub const fn get_latest() -> Self

pub const fn current_nightly() -> Self

pub fn get_compiler_toolchain(&self) -> &'static str

pub fn get_directory_name(&self) -> &'static str