Function wasi_test_generator::wasitests::compile_wasm_for_version

fn compile_wasm_for_version(
    temp_dir: &Path,
    file: &str,
    out_dir: &Path,
    rs_mod_name: &str,
    version: WasiVersion,
) -> Result<PathBuf>

compile the Wasm file for the given version of WASI

returns the path of where the wasm file is