Crate wasi_test_generator

Modules

set_up_toolchain

util

wasi_version

wasitests
This file will run at build time to autogenerate the WASI regression tests It will compile the files indicated in TESTS, to:executable and .wasm

Structs

TestGenOptions

WasiOptions
The options provied when executed a WASI Wasm program

WasiTest
This is the structure of the .wast file

Enums

WasiVersion

Statics

ALL_WASI_VERSIONS

LATEST_WASI_VERSION

Functions

build

install_toolchains

main