wasi_test_generator

Struct TestGenOptions

pub struct TestGenOptions {
    pub(crate) free: Vec<String>,
    pub(crate) nightly: bool,
    pub(crate) all_versions: bool,
    pub(crate) generate_wasm: bool,
    pub(crate) set_up_toolchain: bool,
    pub(crate) help: bool,
}

Fields
free: Vec<String>

if you want to specify specific tests to generate

nightly: bool

Whether to use the current nightly instead of the latest snapshot0 compiler

all_versions: bool

Whether or not to do operations for all versions of WASI or just the latest.

generate_wasm: bool

Whether or not the Wasm will be generated.

set_up_toolchain: bool

Whether or not the logic to install the needed Rust compilers is run.

help: bool

Print the help message