Struct wasi_test_generator::TestGenOptions

pub struct TestGenOptions {
    free: Vec<String>,
    all_versions: bool,
    generate_wasm: bool,
    set_up_toolchain: bool,
    help: bool,
}

Fields
free: Vec<String>

if you want to specify specific tests to generate

all_versions: bool

Whether or not to do operations for all versions of WASI or just the latest.

generate_wasm: bool

Whether or not the Wasm will be generated.

set_up_toolchain: bool

Whether or not the logic to install the needed Rust compilers is run.

help: bool

Print the help message