Function wasi_test_generator::wasitests::generate_native_output

fn generate_native_output(
    temp_dir: &Path, 
    file: &str, 
    normalized_name: &str, 
    args: &[String], 
    options: &WasiOptions
) -> Result<NativeOutput>

Compile and execute the test file as native code, saving the results to be compared against later.

This function attempts to clean up its output after it executes it.